requirements.txt
# Python 3.10.6
datasets
Pillow
imantics
